tests/wpt/web-platform-tests/dom/nodes/CharacterData-appendData.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>CharacterData.appendData</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-appenddata">
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-data">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function testNode(node) {
  test(function() {
    assert_equals(node.data, "test")
    node.appendData("test")
    assert_equals(node.data, "testtest")
  })
}
test(function() {
  testNode(document.createTextNode("test"))
  testNode(document.createComment("test"))
})
</script>
